AT-152: move MaterialModule to new shared
