+++ frontend/src/app/core/main/toolbar/toolbar.module.ts
@@ -1,7 +1,7 @@
 import { NgModule } from '@angular/core';
 import { CommonModule } from '@angular/common';
 import { ToolbarComponent } from './toolbar.component';
-import { MaterialModule } from '../../../modules/material/material.module';
+import { MaterialModule } from '../../../shared/modules/material/material.module';
 import { RouterModule } from '@angular/router';
 
 @NgModule({
