+++ frontend/src/app/feature/jobs/jobs.module.ts
@@ -4,7 +4,7 @@ import { CommonModule } from '@angular/common';
 import { JobsRoutingModule } from './jobs-routing.module';
 import { JobsComponent } from './jobs.component';
 import { SearchModule } from './search/search.module';
-import { MaterialModule } from '../../modules/material/material.module';
+import { MaterialModule } from '../../shared/modules/material/material.module';
 import { JobCardComponent } from './job-card/job-card.component';
 import { JobDetailsComponent } from './job-details/job-details.component';
 import { JobListComponent } from './job-list/job-list.component';
