+++ frontend/src/app/feature/jobs/search/search.module.ts
@@ -1,7 +1,7 @@
 import { NgModule } from '@angular/core';
 import { CommonModule } from '@angular/common';
 import { SearchComponent } from './search.component';
-import { MaterialModule } from '../../../modules/material/material.module';
+import { MaterialModule } from '../../../shared/modules/material/material.module';
 import { ReactiveFormsModule } from '@angular/forms';
 
 @NgModule({
